chore(i18n): stop reporting parser-facing literals as untranslated copy
The i18n gate reports 76 template literals that are handed to a PARSER, never to
a person. Each is exempted here by WHERE the string lives, per this config's own
doctrine, and each was MEASURED for blast radius rather than argued for.

- src/lib/widgetSrcdoc.ts, src/lib/mcpAppSrcdoc.ts (exact paths): CSP directives,
  DOCTYPE, <meta>/<style>/<script> markup, the frame reset CSS, two JS bootstraps.
  Verified copy-free: neither imports i18nT, neither renders a text node.
- ^log(Debug|Warn)$ (callee): popoutController's two console shims. A CALLEE and
  deliberately not a whole-file exemption, because that module also raises a
  translated window.alert -- releasing the file would release real copy.
- src/hooks/themeCss.ts (new): useTheme.tsx's 9 CSS builders could not be exempted
  in place, because the same file declares THEMES, 19 user-visible picker labels.
  Extracting the CSS makes the boundary a path again. THEMES stays in useTheme.tsx
  and stays gated.

Measured: 1408 -> 1332 findings, and only the four target files move
(widgetSrcdoc 39->0, mcpAppSrcdoc 16->0, popoutController 5->1, useTheme 34->17).
No unrelated file changes count, which is what the doctrine actually requires.

Two candidates were REJECTED under the same rule rather than shipped: a
'^rgba?\(' content regex (it leaked into apps/issue-radar/lib/format.ts, a file
outside the target set) and a whole-file release of popoutController.ts.
