+++ website/eslint.i18n.config.js
@@ -99,6 +99,30 @@ export default [
       // false positive it fixes. A path this narrow cannot exempt a future file
       // that does hold copy.
       'src/apps/meetings/lib/sketchSrcdoc.ts',
+      // The widget and MCP-app srcdoc builders — the same category as
+      // `sketchSrcdoc.ts` directly above, and listed by the same exact-path rule
+      // rather than a shared `*Srcdoc.ts` glob. Every literal in both is handed to
+      // a PARSER: CSP directives, a DOCTYPE, `<meta>`/`<style>`/`<script>` markup,
+      // the frame's own reset CSS, and two fixed JS bootstraps (the height reporter
+      // and the nav beacon). Translating any of them would not change a word anyone
+      // reads — it would break the policy or blank the frame.
+      //
+      // Verified copy-free rather than assumed: neither module imports `i18nT` /
+      // `useTranslation`, neither renders a text node, and their exports are pure
+      // builders. The consumers that DO show copy (`WidgetFrame.tsx`, the MCP app
+      // host) stay fully gated.
+      //
+      // Two exact paths, not a content regex. The `^(default|script|img|…)-src\b`
+      // exclusion that the `sketchSrcdoc.ts` note above records as tried-and-rejected
+      // is exactly what would cover these, and it leaked into unrelated files; a
+      // measured `^rgba?\(` variant leaked into `apps/issue-radar/lib/format.ts`.
+      // A path this narrow cannot leak.
+      //
+      // Stated as a false-negative class, per this file's convention: any
+      // user-visible copy ever added to these two paths will not be reported —
+      // keep both modules parser-facing only.
+      'src/lib/widgetSrcdoc.ts',
+      'src/lib/mcpAppSrcdoc.ts',
       // The PPTX Maker board-preview builder — the same category as
       // `sketchSrcdoc.ts` directly above, and listed by the same exact-path rule
       // rather than a shared glob. Every literal in it is handed to a PARSER: the
@@ -139,6 +163,35 @@ export default [
       'src/apps/mochi/src/shared/shortcut.ts',
       // CSS text injected through <style>; a stylesheet is not translatable copy.
       'src/apps/spec-builder/inlineStyles.ts',
+      // The theme stylesheet builders, extracted out of `hooks/useTheme.tsx` so
+      // they COULD be exempted by path. Every literal in the module is handed to
+      // the CSS parser: the `--*` custom-property allowlist, the
+      // `[data-theme="custom-<slug>-<mode>"]` selectors, the static font/radius/
+      // color-scheme defaults, `@font-face` blocks, and the `url()` values the
+      // overrides-pipeline rewrites to pack asset routes. Translating any of them
+      // would not change a word anyone reads — it would unstyle the theme or
+      // silently weaken the §4.2/§5.1 selector scoper.
+      //
+      // Verified copy-free rather than assumed, and by a MECHANICAL boundary
+      // rather than a reading: the module imports no `i18nT` / `useTranslation`,
+      // and it does not touch the DOM at all — every export is data in, CSS
+      // `string` out. The `<style>` tags, `document.head` writes and
+      // `style.setProperty` calls all stayed in `useTheme.tsx`, so nothing here
+      // has a path to the screen.
+      //
+      // This split is the whole reason a path exemption is admissible here.
+      // `useTheme.tsx` also declares `THEMES` — 19 theme-picker display names —
+      // so exempting THAT file would hand back real copy, which is what the
+      // `object-properties: next` note above refuses. Extracting the CSS leaves
+      // the 19 names in a still-gated file. Measured: `useTheme.tsx` 34 -> 17
+      // (exactly the 17 CSS findings), the new module contributes 0, and no other
+      // file's count moves.
+      //
+      // Stated as a false-negative class, per this file's convention: any
+      // user-visible copy ever added to THIS path will not be reported — keep the
+      // module parser-facing only, and keep it DOM-free, which is the property
+      // that makes that easy to check.
+      'src/hooks/themeCss.ts',
     ],
     linterOptions: {
       // Every `eslint-disable` comment in this codebase targets the MAIN config's
@@ -411,6 +464,21 @@ export default [
             exclude: [
               // Diagnostics and dev-only output.
               '^console\\.\\w+$', '^(Type)?Error$', '^URL(SearchParams)?$',
+              // `popoutController.ts`'s two console shims: `logDebug` is
+              // `console.debug` and `logWarn` is `console.warn`, both behind a debug
+              // flag. Identical class to `^console\.\w+$` one line up — the argument
+              // is operator diagnostics ("direct focus of … vetoed"), never a string
+              // the UI renders.
+              //
+              // A CALLEE exemption and deliberately NOT a whole-file one: the module
+              // also raises a `window.alert(i18nT(…))`, so releasing the file would
+              // release a module that does render copy — it would fail the
+              // "verified copy-free" standard the exact-path precedents above meet.
+              //
+              // Known false negative, stated: a future `logDebug`/`logWarn` defined
+              // in another module inherits this. Both names exist in exactly one
+              // module today (`src/utils/popoutController.ts`).
+              '^log(Debug|Warn)$',
               // Validator diagnostics, for parity with `Error` above. A rejected input's
               // reason names the FIELD that failed (`Missing or invalid "meta" field`,
               // `Invalid meta.format: "…" (expected "svg", "lottie", or "sprite")`) and
